requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.8.2
captcha==0.5.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==41.0.4
cycler==0.12.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.5
django-allauth==0.57.0
filelock==3.12.2
Flask==3.0.3
fonttools==4.53.0
h11==0.14.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mysqlclient==2.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.82
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
Pillow==10.0.0
platformdirs==3.10.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
razorpay==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
schedule==1.2.2
seaborn==0.13.2
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.5
virtualenv==20.24.3
Werkzeug==3.0.3
wsproto==1.2.0
